Move TicTacToe#start logic to Game class
