+++ bin/tic_tac_toe.rb
@@ -4,8 +4,6 @@ require_relative '../lib/cli'
 require_relative '../lib/tic_tac_toe'
 
 cli = CLI.new
-board = Board.new
-player = Player.new
 
-tic_tac_toe = TicTacToe.new(cli, board, player)
+tic_tac_toe = TicTacToe.new(cli)
 tic_tac_toe.start
