+++ lib/tic_tac_toe.rb
@@ -3,22 +3,12 @@ require_relative './board'
 require_relative './player'
 
 class TicTacToe
-  def initialize(user_interface, board, player)
+  def initialize(user_interface)
     @user_interface = user_interface
-    @board = board
-    @player_one = player
   end
 
   def start
     display_welcome
-    display_board
-
-    display_move_instruction
-
-    position = @player_one.get_move
-    @player_one.make_move(@board, position)
-
-    display_board
   end
 
   private
@@ -26,14 +16,5 @@ class TicTacToe
   def display_welcome
     welcome_message = 'Welcome to a game of Tic-Tac-Toe!'
     @user_interface.display_message welcome_message
-  end
-
-  def display_board
-    @user_interface.display_board @board.state
-  end
-
-  def display_move_instruction
-    instruction = 'Enter a number to make a move in the corresponding square:'
-    @user_interface.display_message instruction
   end
 end
